Update and expand tests and fix bugs.
