+++ src/metapopulation.jl
@@ -9,9 +9,11 @@ such as those that are geographically isolated from one another.
 A meta population can be used in all instances where an ordinary population is
 valid.
 """
-type MetaPopulation
-  populations::Vector{Population}
+type MetaPopulation{T <: Landscape}
+  populations::Vector{Population{T}}
 end
+
+MetaPopulation{T <: Landscape}(mp::MetaPopulation{T}) = MetaPopulation([Population(p) for p = mp.populations])
 
 @doc """rand(::Type{MetaPopulation}, ls::Landscape, popsize::Int64, popct::Int64)
