+++ src/selection/proportional.jl
@@ -2,7 +2,7 @@ import Base.Random: rand
 
 export propsel, propsel!
 
-@doc """propsel(p::Population, ls::Landscape)
+@doc """propsel(p::Population)
 
 Create a new population through proportional selection on the given
 population. Employs the algorithm describe in the Lipowski and
@@ -16,24 +16,24 @@ A. Lipowski and D. Lipowska, “Roulette-wheel selection via stochastic
 acceptance,” Phys. A Stat. Mech. its Appl., vol. 391, no. 6, pp. 2193–2196,
 2012.
 """
-function propsel(p::Population, ls::Landscape)
-  np = copy(p)
-  propsel!(np, ls)
+function propsel(p::Population)
+  np = Population(p)
+  propsel!(np)
   return np
 end
 
-function propsel(p::MetaPopulation, ls::Landscape)
-  np = copy(p)
-  propsel!(np, ls)
+function propsel(mp::MetaPopulation)
+  np = MetaPopulation(mp)
+  propsel!(np)
   return np
 end
 
-@doc """propsel!(p::Population, ls::Landscape)
+@doc """propsel!(p::Population)
 
 Conduct proportional selection in-place.
 """
-function propsel!(p::Population, ls::Landscape)
-  fs = popfits(p, ls)
+function propsel!(p::Population)
+  fs = popfits(p)
   fmax = maximum(fs)
   if fmax == 0
     # all elements have fitness zero
@@ -52,11 +52,11 @@ function propsel!(p::Population, ls::Landscape)
     end
   end
 
-  p[:,:] = p[:,selected]
+  p.genotypes[:] = p.genotypes[selected]
 end
 
-function propsel!(p::MetaPopulation, ls::Landscape)
-  for ip = 1:popct(p)
-    propsel!(p[:,:,ip], ls)
+function propsel!(mp::MetaPopulation)
+  for p = mp.populations
+    propsel!(p)
   end
 end
