+++ test/runtests.jl
@@ -1,10 +1,4 @@
 # Test runner
 
-include("landscape.jl")
-include("genotype.jl")
-include("neighbors.jl")
-include("walks.jl")
-include("population.jl")
-include("metapopulation.jl")
-include("mutation.jl")
+include("unit.jl")
